Refactor host configurations and improve Bluetooth support
This commit includes several changes to the host configurations for
crambot, lapbot, ninjabot, and soundbot. The use of certain kernels and
the installation of Pulsar and VSCode have been adjusted.

The 'install' and 'repo' sections have been restructured for clarity.
Bluetooth support has been improved with the addition of new packages
and configurations.

The PulseAudio configuration has been updated to include Bluetooth
support, and the installation of Bluetooth packages has been added.

The pacman configuration has been adjusted to accommodate different
architectures.

Finally, the main variable file has been updated to include a new
'use_bluetooth' flag.

This commit improves the flexibility and functionality of the system
configuration.